requirements.txt
langchain
langchain-community
langchain-openai
langgraph
python-dotenv
duckduckgo-search
wikipedia
